The indexlump was converted to uint32_t from float. Which was Causing the Crash

+++ Pehen/vertex_menagerie.h
@@ -26,6 +26,6 @@ private:
 	int indexOffset;
 	vk::Device logicalDevice;
 	std::vector<float> vertexlump;
-	std::vector<float> indexlump;
+	std::vector<uint32_t> indexlump;
 
 };
